Moved zeroconf logic to frameworks
