+++ src/freeseer/framework/announcer.py
